// blog-frontend/src/api/generated/model/pagedResponsePostCardResponse.ts
/**
 * Generated by orval v8.0.0-rc.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PostCardResponse } from './postCardResponse';

export interface PagedResponsePostCardResponse {
  content?: PostCardResponse[];
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
  isLast?: boolean;
}